class OffersList
